NHSO-21907: Receipt report integration tests
