+++ tests/Nhs.App.Api.Integration.Tests/Nhs.App.Api.Integration.Tests/PatientReportRetrievalTests.cs
@@ -7,14 +7,13 @@ using FluentAssertions;
 using Hl7.Fhir.Model;
 using Hl7.Fhir.Serialization;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json;
 using Nhs.App.Api.Integration.Tests.Extensions;
 using Task = System.Threading.Tasks.Task;
 
 namespace Nhs.App.Api.Integration.Tests
 {
     [TestClass]
-    public class EventReportsHttpFunctionsTests : CommunicationHttpFunctionBase
+    public class PatientReportRetrievalTests : HttpEndpointTestsBase
     {
         private static TestConfiguration _testConfiguration;
 
@@ -26,14 +25,59 @@ namespace Nhs.App.Api.Integration.Tests
         }
 
         [TestMethod]
-        public async Task EventReportGet_NoFileAvailableForRequestedDate_Returns404NotFound()
+        public async Task PatientReportGet_InvalidBearerToken_Returns401Unauthorized()
+        {
+            // Arrange
+            using var httpClient = CreateHttpClient();
+
+            var correlationId = Guid.NewGuid().ToString();
+
+            // Act
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=A12345&page=1", correlationId, "invalidAccessToken");
+
+            // Assert
+            response.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
+
+            var operationOutcome = await ParseOperationOutcome(response);
+
+            var issue = operationOutcome.Issue.Single();
+            issue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
+            issue.Code.Should().Be(OperationOutcome.IssueType.Forbidden);
+            issue.Diagnostics.Should().Be("Invalid Access Token");
+
+            response.Headers.ShouldContainHeader("X-Correlation-ID", correlationId);
+        }
+
+        [TestMethod]
+        public async Task PatientReportGet_BlankOdsCodeParameter_Returns400BadRequest()
+        {
+            // Arrange
+            using var httpClient = CreateHttpClient();
+
+            // Act
+            var response = await httpClient.GetAsync("communication/report/patients/?ods-organisation-code=&page=1");
+
+            // Assert
+            response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
+
+            var operationOutcome = await ParseOperationOutcome(response);
+
+            var issue = operationOutcome.Issue.Single();
+            issue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
+            issue.Code.Should().Be(OperationOutcome.IssueType.Invalid);
+            issue.Diagnostics.Should().Contain("Invalid format");
+            issue.Expression.Single().Should().Be("ods-organisation-code");
+        }
+
+        [TestMethod]
+        public async Task PatientReportGet_NoFileAvailableForRequestedOdsCode_Returns404NotFound()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
             var response =
-                await httpClient.GetAsync($"communication/report/events/?day=2001-01-01&page=1");
+                await httpClient.GetAsync($"communication/report/patients/?ods-organisation-code=C12345&page=1");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.NotFound);
@@ -47,14 +91,14 @@ namespace Nhs.App.Api.Integration.Tests
         }
 
         [TestMethod]
-        public async Task EventReportGet_MultipleDaysRequested_Returns400BadRequest()
+        public async Task PatientReportGet_MultipleOdsCodesRequested_Returns400BadRequest()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
             var response =
-                await httpClient.GetAsync("communication/report/events/?day=2021-03-01&day=2021-03-02&page=1");
+                await httpClient.GetAsync("communication/report/patients/?ods-organisation-code=A12345&ods-organisation-code=A12355&page=1");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -64,18 +108,18 @@ namespace Nhs.App.Api.Integration.Tests
             var issue = operationOutcome.Issue.Single();
             issue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
             issue.Code.Should().Be(OperationOutcome.IssueType.Invalid);
-            issue.Expression.Single().Should().Be("day");
+            issue.Expression.Single().Should().Be("ods-organisation-code");
             issue.Diagnostics.Should().Be("A single value is expected");
         }
 
         [TestMethod]
-        public async Task EventReportGet_MultiplePagesRequested_Returns400BadRequest()
+        public async Task PatientReportGet_MultiplePagesRequested_Returns400BadRequest()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
-            var response = await httpClient.GetAsync("communication/report/events/?day=2021-03-01&page=1&page=2");
+            var response = await httpClient.GetAsync("communication/report/patients/?ods-organisation-code=A12345&page=1&page=2");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -89,22 +133,15 @@ namespace Nhs.App.Api.Integration.Tests
             issue.Diagnostics.Should().Be("A single value is expected");
         }
 
-        [DataTestMethod]
-        [DataRow("2021-02-30", "1")]
-        [DataRow("2021-13-01", "1")]
-        [DataRow("09-03-2021", "1")]
-        [DataRow("2021-3-9", "1")]
-        [DataRow("2021-03-08T01:00:00", "1")]
-        [DataRow("yesterday", "1")]
-        public async Task EventReportGet_InvalidDayParameter_Returns400BadRequest(string dayParameter,
-            string pageParameter)
+        [TestMethod]
+        public async Task PatientReportGet_InvalidOdsCodeParameter_Returns400BadRequest()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
             var response =
-                await httpClient.GetAsync($"communication/report/events/?day={dayParameter}&page={pageParameter}");
+                await httpClient.GetAsync($"communication/report/patients/?ods-organisation-code=C12345C&page=1");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -114,78 +151,44 @@ namespace Nhs.App.Api.Integration.Tests
             var issue = operationOutcome.Issue.Single();
             issue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
             issue.Code.Should().Be(OperationOutcome.IssueType.Invalid);
-            issue.Expression.Single().Should().Be("day");
+            issue.Expression.Single().Should().Be("ods-organisation-code");
             issue.Diagnostics.Should().Be("Invalid format");
         }
 
-        [DataTestMethod]
-        [DataRow("2021-01-01", "1", @"{""Name"": ""Integration test file 1""}")]
-        [DataRow("2021-01-02", "2", @"{""Name"": ""Integration test file 2""}")]
-        public async Task EventReportGet_ValidRequestForDayWhereFileExists_ContentsOfFileAreReturned(
-            string dayParameter, string pageParameter, string expectedFileContents)
+        [TestMethod]
+        public async Task PatientReportGet_ValidRequestForOdsCodeWhereSinglePageExists_ContentsAndLinkHeaderReturned()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
-            var path = $"communication/report/events?day={dayParameter}&page={pageParameter}";
 
             // Act
-            var response = await httpClient.GetAsync(path);
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=A12345&page=1");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.OK);
 
             var responseString = await response.Content.ReadAsStringAsync();
-            responseString.Should().Be(expectedFileContents);
-            response.Headers.GetValues("Link").Single().Should().BeEquivalentTo($"<{httpClient.BaseAddress}{path}>; rel=\"last\"");
+            responseString.Should().Be(@"[{""NhsNumber"": ""987654321"", ""NotificationsEnabled"": ""true""}]");
+            response.Headers.GetValues("Link").Single().Should().BeEquivalentTo($"<{httpClient.BaseAddress}{$"communication/report/patients?ods-organisation-code=A12345&page=1"}>; rel=\"last\"");
         }
 
         [TestMethod]
-        public async Task EventReportGet_NoCorrelationIdPassed_NoCorrelationIdHeaderInTheResponse()
+        public async Task PatientReportGet_ValidRequestForOdsCodeWhereMultiplePagesExist_ContentsAndLinkHeaderReturned()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
-            var response = await httpClient.GetAsync("communication/report/events/?day=2021-03-08");
-
-            // Assert
-            response.Headers.ShouldNotContainHeader("X-Correlation-ID");
-        }
-
-        [TestMethod]
-        public async Task EventReportGet_ValidRequestForDayWhereSinglePageExists_ContentsAndLinkHeaderReturned()
-        {
-            // Arrange
-            using var httpClient = CreateHttpClient();
-
-            // Act
-            var response = await httpClient.GetAsync("communication/report/events?day=2021-01-01&page=1");
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=B12345&page=1");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.OK);
 
             var responseString = await response.Content.ReadAsStringAsync();
-            responseString.Should().Be(@"{""Name"": ""Integration test file 1""}");
-            response.Headers.GetValues("Link").Single().Should().BeEquivalentTo($"<{httpClient.BaseAddress}{$"communication/report/events?day=2021-01-01&page=1"}>; rel=\"last\"");
-        }
-
-        [TestMethod]
-        public async Task EventReportGet_ValidRequestForDayWhereMultiplePagesExist_ContentsAndLinkHeaderReturned()
-        {
-            // Arrange
-            using var httpClient = CreateHttpClient();
-
-            // Act
-            var response = await httpClient.GetAsync("communication/report/events?day=2021-01-02&page=1");
-
-            // Assert
-            response.StatusCode.Should().Be(HttpStatusCode.OK);
-
-            var responseString = await response.Content.ReadAsStringAsync();
-            responseString.Should().Be(@"{""Name"": ""Integration test file 2""}");
+            responseString.Should().Be(@"[{""NhsNumber"": ""987654321"", ""NotificationsEnabled"": ""true""}]");
             var expectedLinkHeader =
-                $"<{httpClient.BaseAddress}{"communication/report/events?day=2021-01-02&page=2"}>; rel=\"next\", " +
-                $"<{httpClient.BaseAddress}{"communication/report/events?day=2021-01-02&page=2"}>; rel=\"last\"";
+                $"<{httpClient.BaseAddress}{"communication/report/patients?ods-organisation-code=B12345&page=2"}>; rel=\"next\", " +
+                $"<{httpClient.BaseAddress}{"communication/report/patients?ods-organisation-code=B12345&page=2"}>; rel=\"last\"";
             response.Headers.GetValues("Link").Single().Should().BeEquivalentTo(expectedLinkHeader);
         }
 
@@ -195,14 +198,14 @@ namespace Nhs.App.Api.Integration.Tests
         [DataRow("0")]
         [DataRow("2147483648")]
         [DataRow("")]
-        public async Task EventReportGet_InvalidRequestInvalidPageValue_ReturnsBadRequest(string pageParameter)
+        public async Task PatientReportGet_InvalidRequestInvalidPageValue_ReturnsBadRequest(string pageParameter)
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
             var response =
-                await httpClient.GetAsync($"communication/report/events?day=2021-01-01&page={pageParameter}");
+                await httpClient.GetAsync($"communication/report/patients?ods-organisation-code=A12345&page={pageParameter}");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -217,14 +220,14 @@ namespace Nhs.App.Api.Integration.Tests
         }
 
         [TestMethod]
-        public async Task EventReportGet_InvalidRequestNoPageValue_ReturnsBadRequest()
+        public async Task PatientReportGet_InvalidRequestNoPageValue_ReturnsBadRequest()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
             var response =
-                await httpClient.GetAsync($"communication/report/events?day=2021-01-01&page=");
+                await httpClient.GetAsync($"communication/report/patients?ods-organisation-code=A12345&page=");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -239,26 +242,26 @@ namespace Nhs.App.Api.Integration.Tests
         }
 
         [TestMethod]
-        public async Task EventReportGet_ValidRequestNoPageParameter_ReturnsOk()
+        public async Task PatientReportGet_ValidRequestNoPageParameter_ReturnsOk()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
-            var response = await httpClient.GetAsync("communication/report/events?day=2021-01-01");
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=A12345");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.OK);
         }
 
         [TestMethod]
-        public async Task EventReportGet_InvalidPageAndDateParameters_ReturnsBadRequest()
+        public async Task PatientReportGet_InvalidPageAndOdsCodeParameters_ReturnsBadRequest()
         {
             // Arrange
             using var httpClient = CreateHttpClient();
 
             // Act
-            var response = await httpClient.GetAsync("communication/report/events?day=abc&page=abc");
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=abc&page=abc");
 
             // Assert
             response.StatusCode.Should().Be(HttpStatusCode.BadRequest);
@@ -266,12 +269,12 @@ namespace Nhs.App.Api.Integration.Tests
             var operationOutcome = await ParseOperationOutcome(response);
             operationOutcome.Issue.Count.Should().Be(2);
 
-            var dayError = operationOutcome.Issue.Where(i => i.ExpressionElement.Where(e => e.Value == "day").Any());
-            var dayIssue = dayError.Single();
-            dayIssue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
-            dayIssue.Code.Should().Be(OperationOutcome.IssueType.Invalid);
-            dayIssue.Expression.Single().Should().Be("day");
-            dayIssue.Diagnostics.Should().Be("Invalid format");
+            var odsCodeError = operationOutcome.Issue.Where(i => i.ExpressionElement.Where(e => e.Value == "ods-organisation-code").Any());
+            var odsCodeIssue = odsCodeError.Single();
+            odsCodeIssue.Severity.Should().Be(OperationOutcome.IssueSeverity.Error);
+            odsCodeIssue.Code.Should().Be(OperationOutcome.IssueType.Invalid);
+            odsCodeIssue.Expression.Single().Should().Be("ods-organisation-code");
+            odsCodeIssue.Diagnostics.Should().Be("Invalid format");
 
             var pageError = operationOutcome.Issue.Where(i => i.ExpressionElement.Where(e => e.Value == "page").Any());
             var pageIssue = pageError.Single();
@@ -290,7 +293,7 @@ namespace Nhs.App.Api.Integration.Tests
             var correlationId = Guid.NewGuid().ToString();
 
             // Act
-            var response = await httpClient.GetAsync("communication/report/events/?day=2021-03-08",
+            var response = await httpClient.GetAsync("communication/report/patients?ods-organisation-code=A12345",
                 correlationId, "invalidAccessToken");
 
             // Assert
